New chaosuploader
We separate the logic to upload to chaos to preserve the SRP away from utils.
